Move CdbXMLToJsonLDLabelImporter to Cdb namespace
